user_service/requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
mako==1.3.6
markupsafe==3.0.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
py-env==0.0.1
pyjwt==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.0.6
